requirements.txt
Django==4.0.5
psycopg2==2.9.3
openpyxl==3.0.10
djangorestframework==3.13.1
drf-yasg==1.20.0